implemented BSP tree to mood

+++ editor/src/vector2.cpp
@@ -1,5 +1,5 @@
 #include "vector2.h"
 
 std::ostream &operator<<(std::ostream &os, Vector2 const &m) {
-	return os << (int)m.x << "," << (int)m.y;
+	return os << m.x << "," << m.y;
 }
